+++ apple/tun2proxy/Tun2proxyWrapper.m
@@ -14,7 +14,7 @@
 
 + (void)startWithConfig:(NSString *)proxy_url
                  tun_fd:(int)tun_fd
-                tun_mtu:(uint32_t)tun_mtu
+                tun_mtu:(uint16_t)tun_mtu
            dns_over_tcp:(bool)dns_over_tcp
                 verbose:(bool)verbose {
   ArgDns dns_strategy = dns_over_tcp ? OverTcp : Direct;
